dev-requirements.txt
colorama==0.3.9
flake8==3.5.0
mccabe==0.6.1
mypy==0.630
mypy-extensions==0.4.1
pycodestyle==2.3.1
pyflakes==1.6.0
toml==0.9.6
typed-ast==1.1.0
